install

+++ linux/terminator/install_terminator_config.sh
@@ -1,8 +1,16 @@
 #! /bin/bash
 
-if [ ! -L $HOME/.config/terminator/config ]
+if [ ! -e $HOME/config/terminator ]
 then
-	mv $HOME/.config/terminator/config $HOME/.config/terminator/config.old
-	ln -s $(pwd)/config $HOME/.config/terminator/config
+    mkdir $HOME/.config/terminator
 fi
 
+if [ -e  $HOME/.config/terminator/config ] && [ ! -L $HOME/.config/terminator/config ]
+then
+	mv $HOME/.config/terminator/config $HOME/.config/terminator/config.old
+fi
+
+if [ ! -L $HOME/.config/terminator/config ]
+then
+    ln -s $(pwd)/config $HOME/.config/terminator/config
+fi
